Remove getCommunityCareFacility and move mockCCProviderFetch
