+++ src/applications/vaos/services/var/index.js
@@ -1,4 +1,4 @@
-import { apiRequestWithUrl, parseApiList, parseApiObject } from '../utils';
+import { apiRequestWithUrl, parseApiList } from '../utils';
 
 export async function getCommunityCareFacilities({
   latitude,
@@ -15,10 +15,4 @@ export async function getCommunityCareFacilities({
   return apiRequestWithUrl(
     `/facilities_api/v1/ccp/provider?latitude=${latitude}&longitude=${longitude}&radius=${radius}&per_page=${perPage}&page=${page}&${bboxQuery}&${specialtiesQuery}&trim=true`,
   ).then(parseApiList);
-}
-
-export async function getCommunityCareFacility(id) {
-  return apiRequestWithUrl(`/v1/facilities/ccp/${id}`, {
-    method: 'GET',
-  }).then(parseApiObject);
 }
