+++ src/applications/vaos/services/var/tests/index.unit.spec.js
@@ -4,7 +4,7 @@ import {
   setFetchJSONResponse,
 } from '@department-of-veterans-affairs/platform-testing/helpers';
 import environment from '@department-of-veterans-affairs/platform-utilities/environment';
-import { getCommunityCareFacilities, getCommunityCareFacility } from '..';
+import { getCommunityCareFacilities } from '..';
 import { calculateBoundingBox } from '../../../utils/address';
 
 describe('VAOS Services: CC facilities API', () => {
@@ -17,24 +17,6 @@ describe('VAOS Services: CC facilities API', () => {
       attributes: {},
     },
   ];
-
-  it('should get community care facility', async () => {
-    // Arrange
-    const id = 1;
-
-    mockFetch();
-    setFetchJSONResponse(global.fetch, { data });
-
-    // Act
-    await getCommunityCareFacility(id);
-
-    // Assert
-    Sinon.assert.calledOnce(global.fetch);
-    Sinon.assert.calledWith(
-      global.fetch,
-      `${environment.API_URL}/v1/facilities/ccp/${id}`,
-    );
-  });
 
   it('should get community care facilities', async () => {
     // Arrange
